Refactor: Image grid visualizer
The logic for visualizing mazes as images now works for both modified
and unmodified grids.

The image visualizer now uses a builder to set optional parameters such
as the cell size, border size, background color, and grid color. Also, a
directory is created if it does not already exist for the specified
image path.

Resolves: #PMG-14